fsda sadase

+++ Models/PostQuotaDistribution.cs
@@ -8,8 +8,8 @@ namespace ResultManager.Models
         public int Id { get; set; }
         public string PostName { get; set; }
         public string QuotaName { get; set; }
-        public float DecimalQuantity { get; set; }
-        public float RoundedQuantity { get; set; }
+        public double DecimalQuantity { get; set; }
+        public double RoundedQuantity { get; set; }
         public int ApplicantFound { get; set; }
     }
 }
