+++ Models/Quotas.cs
@@ -7,7 +7,7 @@ namespace ResultManager.Models
     {
         public int Id { get; set; }
         public string Name { get; set; }
-        public float Percentage { get; set; }
+        public double Percentage { get; set; }
         public int Priority { get; set; }
     }
 }
